Refactored for testing #11
Added a test library and refactored the app so it all fits into `npm
test` nicely.
