+++ fixtures/workouts.yaml
@@ -1,0 +1,20 @@
+fixtures:
+  - model: User
+    data:
+      id: 1
+      email: test@example.com
+      password: password
+      first_name: Test
+      last_name: User
+
+  - model: Location
+    data:
+      id: 1
+      name: Test Location
+
+  - model: Workout
+    data:
+      id: 1
+      userId: 1
+      locationId: 1
+      workout_date: 2016-01-10
