--- test/fixtures/Exercise.json
@@ -1,14 +1,0 @@
-[
-  {
-    "id": 1,
-    "name": "Bench Press",
-    "location": 1,
-    "machine_number": "123"
-  },
-  {
-    "id": 2,
-    "name": "Squat Machine",
-    "location": 1,
-    "machine_number": "789"
-  }
-]
