--- test/fixtures/Location.json
@@ -1,8 +1,0 @@
-[
-  {
-    "id": 1,
-    "name": "Test Gym",
-    "address_line": "Test Line",
-    "postcode": "SW19 1AB"
-  }
-]
