--- test/fixtures/User.json
@@ -1,16 +1,0 @@
-[
-  {
-    "id": 1,
-    "is_staff": true,
-    "password": "$2a$10$Yp6yN.ouWKM142twziQ6J.lcZzC3mrvy3mQAB85d1Qijp46YwEhEu",
-    "email": "test@example.com",
-    "resetToken": null
-  },
-  {
-    "id": 2,
-    "is_staff": false,
-    "email": "user@example.com",
-    "password": "$2a$10$K8z4iyupILqk2C4c1QklS.BzZzPTQjAy2WrLfGB5p4wDSEiZ3Tl2O",
-    "resetToken": null
-  }
-]
