--- test/fixtures/Workout.json
@@ -1,20 +1,0 @@
-[
-  {
-    "id": 1,
-    "user": 1,
-    "workout_date": "2015-10-10",
-    "location": 1
-  },
-  {
-    "id": 2,
-    "user": 1,
-    "workout_date": "2015-11-10",
-    "location": 1
-  },
-  {
-    "id": 3,
-    "user": 2,
-    "workout_date": "2014-11-10",
-    "location": 1
-  }
-]
